backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0
redis>=5.0.0
celery>=5.3.0
pydantic-settings>=2.0.0
pydantic>=2.0.0
httpx>=0.27.0
python-jose[cryptography]>=3.3.0

# Password hashing — use bcrypt directly (passlib is incompatible with bcrypt >= 4.0)
bcrypt>=4.0.0

python-multipart>=0.0.9
email-validator>=2.1.0
python-dotenv>=1.0.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
razorpay>=1.3.0